requirements.txt
asgiref==3.6.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
cftime==1.6.2
chardet==5.1.0
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.7.0
dbfread==2.0.7
Django==4.1.6
django-cors-headers==3.14.0
django-excel==0.0.10
django-filter==22.1
django-sortedm2m==3.1.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
et-xmlfile==1.1.0
fonttools==4.39.3
freetype-py==2.3.0
future==0.18.3
geojson==3.0.1
geoserver-restconfig==2.0.8
gisdata==0.5.4
html5lib==1.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
json2html==1.3.0
lml==0.1.0
lxml==4.9.2
MarkupSafe==2.1.2
netCDF4==1.6.3
numpy==1.24.2
openpyxl==3.1.2
packaging==23.0
pandas==2.0.1
Pillow==9.4.0
psycopg2==2.9.5
pycairo==1.23.0
pycparser==2.21
pydyf==0.6.0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pygdaltools==1.3.2
PyJWT==2.6.0
pyphen==0.14.0
python-dateutil==2.8.2
pytz==2022.7.1
reportlab==4.0.0
requests==2.28.2
rlPyCairo==0.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
shapely==2.0.1
six==1.16.0
sqlparse==0.4.3
sridentify==0.8.0
tablib==3.4.0
texttable==1.6.7
tinycss2==1.2.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
webencodings==0.5.1
xarray==2023.4.2
XlsxWriter==3.1.0
xmltodict==0.13.0
zopfli==0.2.2
